🐍 Update: Galaxy Update
