+++ Galaxy_Kivy/GalaxyFunctions/Keyboard.py
@@ -1,0 +1,34 @@
+def keyboard_closed(self):
+    self._keyboard.unbind(on_key_down=self._on_keyboard_down)
+    self._keyboard.unbind(on_key_up=self._on_keyboard_up)
+    self._keyboard = None
+
+def on_keyboard_down(self, keyboard, keycode, text, modifiers):
+    if keycode[1] == 'left':
+        self.current_speed_x = self.SPEED_X
+
+    elif keycode[1] == 'right':
+        self.current_speed_x = -self.SPEED_X
+        
+    return True
+
+def on_keyboard_up(self, keyboard, keycode):
+    self.current_speed_x = 0
+    return True
+
+def on_touch_down(self, touch):
+    """
+        Atalhos de botões (mouse) ESQUERDA e DIREITA
+    """
+
+    if touch.x < self.width/2:
+        self.current_speed_x = self.SPEED_X
+
+    else:
+        self.current_speed_x = -self.SPEED_X
+
+def on_touch_up(self, touch):
+    """
+        Atalhos de botões (mouse) CIMA e BAIXO
+    """
+    self.current_speed_x = 0
