+++ Galaxy_Kivy/Galaxy_App.py
@@ -67,6 +67,12 @@ class MainWidget(Widget, MovementApp, PlatformCheck):
         H_LINES_SPACING : espaço entre as linhas
     """
 
+    # Transform.py
+    from GalaxyFunctions.Transform import transform, transform_2D, transform_perspective
+
+    # Keyboard.py
+    from GalaxyFunctions.Keyboard import keyboard_closed, on_keyboard_down, on_keyboard_up, on_touch_down, on_touch_up
+
     perspective_point_x = NumericProperty(0)
     perspective_point_y = NumericProperty(0)
 
@@ -88,81 +94,6 @@ class MainWidget(Widget, MovementApp, PlatformCheck):
             self._keyboard = Window.request_keyboard(self.keyboard_closed, self)
             self._keyboard.bind(on_key_down=self.on_keyboard_down)
             self._keyboard.bind(on_key_up=self.on_keyboard_up)
-
-    def keyboard_closed(self):
-        self._keyboard.unbind(on_key_down=self._on_keyboard_down)
-        self._keyboard.unbind(on_key_up=self._on_keyboard_up)
-        self._keyboard = None
-
-    def on_keyboard_down(self, keyboard, keycode, text, modifiers):
-        if keycode[1] == 'left':
-            self.current_speed_x = self.SPEED_X
-
-        elif keycode[1] == 'right':
-            self.current_speed_x = -self.SPEED_X
-        
-        return True
-
-    def on_keyboard_up(self, keyboard, keycode):
-        self.current_speed_x = 0
-        return True
-
-    def on_touch_down(self, touch):
-        """
-            Atalhos de botões (mouse) ESQUERDA e DIREITA
-        """
-        if touch.x < self.width/2:
-            self.current_speed_x = self.SPEED_X
-
-        else:
-            self.current_speed_x = -self.SPEED_X
-
-    def on_touch_up(self, touch):
-        """
-            Atalhos de botões (mouse) CIMA e BAIXO
-        """
-        self.current_speed_x = 0
-
-    def on_parent(self, widget, parent):
-        # print(f'ON PARENT\nW: {self.width}\nH: {self.height}')
-        pass
-
-    def on_size(self, *args):
-        """
-            Mostra os valores de LARGURA e ALTURA da janela quando alterada.
-        """
-
-        print(f'=====>  ] [ON SIZE     ]\nW: {self.width}\nH: {self.height}\n======')
-        print('')
-
-        # self.perspective_point_x = self.width/2
-        # self.perspective_point_y = self.height * 0.75
-
-    def on_perspective_point_x(self, widget, value):
-        """
-            Mostra apenas o valor de LARGURA(X)
-
-            Args:
-
-                self   : class
-                widget : Janela da aplicação
-                value  : valor de LARGURA
-        """
-
-        print(f'PX: {str(value)}')
-
-    def on_perspective_point_y(self, widget, value):
-        """
-            Mostra apenas o valor de ALTURA(Y)
-
-            Args:
-
-                self   : class
-                widget : Janela da aplicação
-                value  : valor de LARGURA
-        """
-
-        print(f'PY: {str(value)}')
 
     def init_vertical_lines(self):
         """
@@ -231,44 +162,6 @@ class MainWidget(Widget, MovementApp, PlatformCheck):
             x2, y2 = self.transform(xmax, line_y)
 
             self.horizontal_lines[i].points = [x1, y1, x2, y2]
-
-    def transform(self, x, y):
-        """
-            Transformar as linhas para a perspectiva 2D/3D.
-        """
-
-        #return self.transform_2D(x, y)
-        return self.transform_perspective(x, y)
-    
-    def transform_2D(self, x:int, y:int):
-        """
-            Transformar para 2D.
-        """
-
-        return x, y
-    
-    def transform_perspective(self, x:int, y:int):
-        """
-            Criar a perspetiva do cenário.
-        """
-
-        lin_y = y * self.perspective_point_y / self.height
-
-        if lin_y > self.perspective_point_y:
-            lin_y = self.perspective_point_y
-
-        diff_x = x - self.perspective_point_x
-        diff_y = self.perspective_point_y - lin_y
-        factor_y = diff_y / self.perspective_point_y
-        factor_y = pow(factor_y, 2) #número de vezes que será multiplicado
-
-        # TRANSFORM [X]
-        tr_x = self.perspective_point_x + diff_x * factor_y
-
-        # TRANSFORM [Y]
-        tr_y = self.perspective_point_y - factor_y * self.perspective_point_y
-
-        return tr_x, tr_y
 
 class Galaxy(App):
     """
